fix: reduce agentGeneration.ts complexity and line count to pass lint
Move resolvePendingGeneration to agentGenerationRecovery.ts,
runToolOutputGeneration to agentNonStreamGeneration.ts, and
buildKnowledgeTools to agentKnowledge.ts to bring agentGeneration.ts
under the 400-line and complexity-10 lint limits.
